Code Cleaned, commented and removed unused parts
