+++ pom.xml
@@ -75,8 +75,6 @@
                 <artifactId>httpclient</artifactId>
                 <version>4.5.14</version>
             </dependency>
-
-            <!-- https://mvnrepository.com/artifact/org.json/json -->
             <!-- https://mvnrepository.com/artifact/com.google.code.gson/gson -->
             <dependency>
                 <groupId>com.google.code.gson</groupId>
