+++ retorchfiles/envfiles/tjobg.env
@@ -1,0 +1,18 @@
+testsBasePath=./
+tjobname=tjobg
+frontend_port=5048
+retorchportj=5058
+retorchportk=5059
+retorchporth=5056
+mysqlImage=mysql:5.7.21
+retorchporti=5057
+retorchportf=5054
+retorchportg=5055
+retorchportd=5052
+retorchporte=5053
+retorchportb=5050
+retorchportc=5051
+retorchporta=5049
+OpenViduImage=openvidu/openvidu-server-kms:1.7.0
+ESHOP_EXTERNAL_DNS_NAME_OR_IP=$DOCKER_HOST_IP
+ESHOP_STORAGE_CATALOG_URL=http://webshoppingagg_${tjobname}:80/c/api/v1/catalog/items/[0]/pic/
