+++ src/test/java/giis/eshopcontainers/e2e/functional/model/Order.java
@@ -1,4 +1,4 @@
-package giis.eshopcontainers.e2e.api.model;
+package giis.eshopcontainers.e2e.functional.model;
 import java.util.List;
